refactor: Move fixture's to conftest.py
